Constructs the projection direction with fixed tuning parameter in high dimensional regression

Constructs the projection direction with fixed tuning parameter in high dimensional regression

Direction_fixedtuning(
  X,
  loading,
  mu = NULL,
  model = "linear",
  weight = NULL,
  deriv.vec = NULL
)

Arguments

X
Design matrix, of dimension \(n\) x \(p\)

loading
Loading, of length \(p\)

mu
The dual tuning parameter used in the construction of the projection direction

model
The high dimensional regression model, either linear or logistic

weight
The weight vector of length \(n\), used in correcting the plug-in estimator ; to be supplied if model="logistic" (default=NULL)

deriv.vec
The first derivative vector of the logit function at \(X\widehat{\beta}\) ; to be supplied if model="logistic"

Value

proj

The projection direction

Examples

n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
resol = 1.5
step = 3
Direction_fixedtuning(X,loading=c(1,rep(0,(p-1))),mu=sqrt(2.01*log(p)/n)*resol^{-(step-1)})

#> [1] "fixed mu"
#> [1] 0.1542347

#> $proj
#>   [1]  5.332150e-01  2.102406e-06  8.302894e-07  1.129695e-06  3.753662e-07
#>   [6] -1.797747e-06  1.825104e-06  3.128767e-08 -1.048701e-06  2.780636e-06
#>  [11] -1.885179e-06 -1.029269e-06  1.404806e-06  4.395619e-07  3.370703e-06
#>  [16] -4.434226e-06 -2.056283e-06  9.920750e-07 -1.610345e-06 -4.440898e-07
#>  [21]  1.751137e-03  4.339049e-06  6.790858e-08 -2.351764e-06  1.046654e-06
#>  [26] -5.293377e-06 -5.585770e-07 -2.621411e-06 -3.694067e-08 -2.370848e-06
#>  [31] -4.777248e-06  2.011911e-06 -1.777020e-06  1.953230e-06 -1.096476e-03
#>  [36] -4.881362e-08 -2.966551e-06 -3.458888e-07 -2.028550e-06 -9.921419e-07
#>  [41]  9.368670e-03 -2.570106e-06 -8.888999e-07  1.954491e-02 -6.841061e-07
#>  [46] -9.281409e-07  2.105925e-06  6.940905e-06 -1.892885e-06 -3.019416e-02
#>  [51] -1.748448e-06 -4.268960e-06 -9.324041e-07  1.134835e-06 -7.801568e-07
#>  [56] -4.137553e-02  1.618074e-06 -2.366044e-06 -4.065267e-06  3.368630e-02
#>  [61] -1.293848e-06  6.853171e-07  7.773362e-07 -2.594374e-08 -2.769222e-06
#>  [66] -2.950821e-06  1.051263e-06 -8.752512e-07 -4.174753e-06 -1.262254e-07
#>  [71] -5.525017e-06  9.587629e-03  9.249727e-08  1.000082e-06  2.739987e-06
#>  [76]  2.505489e-06 -3.311005e-06 -1.332527e-07 -2.155501e-06 -3.837928e-06
#>  [81] -6.172161e-03 -4.848788e-06  3.140631e-06 -2.914001e-06 -2.937749e-07
#>  [86]  5.262152e-08  2.879321e-06  2.858551e-07  3.231530e-07  8.184142e-07
#>  [91]  4.917296e-07 -4.424169e-06  1.801611e-07 -1.240375e-06  9.852280e-03
#>  [96] -2.884297e-06 -2.368565e-06  2.397489e-06 -1.467353e-06  1.699522e-06
#> [101] -1.415885e-06 -1.778366e-06  5.518067e-06  1.996241e-06  6.341549e-06
#> [106] -2.401634e-06 -1.190794e-06 -8.883273e-07  3.749285e-07 -4.083435e-06
#> [111] -3.708464e-06 -1.931198e-02  4.580428e-07  7.653134e-07 -5.179338e-06
#> [116]  4.356187e-06 -2.962121e-06 -2.272883e-06 -6.463492e-07  2.872640e-06
#> [121] -4.082566e-06 -7.864891e-08 -1.308436e-07 -4.215256e-07  1.861102e-06
#> [126] -2.012477e-06 -2.116322e-06  2.591271e-07  8.752005e-07 -2.099059e-06
#> [131]  3.227737e-02  1.207586e-08  5.027816e-02 -3.676804e-06 -6.337020e-08
#> [136]  2.217729e-06 -9.906264e-07  5.091214e-06 -7.666583e-07 -2.781286e-06
#> [141] -2.318059e-06 -8.595756e-07  2.753414e-06  8.148184e-07 -2.711924e-06
#> [146]  7.120905e-07  9.925227e-07 -1.493474e-06  1.943278e-06 -3.005694e-03
#> [151]  4.510144e-06  1.052259e-06  2.528878e-06 -2.001385e-06 -1.418653e-06
#> [156]  1.449749e-06  5.376032e-07 -5.737604e-06  3.626599e-07 -2.325265e-06
#> [161] -2.158904e-06  1.256843e-06  8.109071e-07 -3.508885e-06  2.363123e-06
#> [166]  6.001416e-07 -7.765799e-02 -5.853209e-07 -2.266220e-06 -2.257753e-06
#> [171]  4.636342e-07 -6.634893e-07 -8.318136e-07  5.891411e-07  1.825176e-06
#> [176]  3.937956e-07  1.659567e-06 -2.899421e-06 -5.231463e-07 -1.144328e-07
#> [181] -6.511361e-07 -1.409529e-06  7.232073e-07 -2.962255e-06 -4.077972e-06
#> [186] -1.500035e-06 -6.985820e-07 -1.779645e-06  1.752719e-06 -2.510419e-06
#> [191] -6.408341e-07 -4.545075e-07  9.225071e-08  2.793603e-06 -1.516556e-02
#> [196] -9.088349e-07  7.113046e-07  1.780670e-06 -4.222109e-07 -2.574033e-06
#> [201]  2.941109e-06  1.478521e-06 -1.847153e-06 -4.421636e-06 -4.030058e-06
#> [206] -6.065794e-06 -4.096578e-06 -3.114912e-06 -3.789333e-06  1.375577e-06
#> [211]  2.980936e-02 -7.089529e-08  2.209202e-06 -7.225394e-03 -2.235048e-06
#> [216] -3.576951e-06 -4.100008e-06 -5.951837e-07 -4.744393e-03  4.368320e-07
#> [221] -8.725731e-07  1.784960e-07  2.945721e-07  1.993370e-06 -1.006679e-02
#> [226]  9.147758e-08  3.724952e-07  7.526453e-07 -6.203097e-02  3.350679e-06
#> [231]  8.078708e-07  1.422211e-06  1.244438e-06  1.675464e-06  2.911419e-06
#> [236]  1.686914e-06  1.870190e-06  1.980641e-06  2.605013e-06 -2.107456e-08
#> [241] -3.259981e-06  2.394079e-06 -2.118058e-03  5.249880e-07 -6.237593e-06
#> [246]  1.707697e-06  7.402575e-07  7.421957e-07 -3.907687e-06  3.438312e-06
#> [251]  2.915648e-06 -6.798090e-06 -2.525884e-06 -1.726517e-06 -9.079182e-07
#> [256] -5.964883e-07  6.351401e-07  4.501230e-06 -3.948710e-06  3.003525e-06
#> [261] -2.463164e-06 -2.836230e-06  5.096427e-06 -2.409177e-06  1.813247e-06
#> [266] -3.317313e-06 -4.441247e-08  3.741744e-06 -3.712036e-06  1.423756e-06
#> [271]  4.928642e-06 -2.566369e-06  7.576949e-07 -3.786697e-07  3.608479e-06
#> [276] -2.875077e-06 -8.443410e-07 -4.519291e-07  4.508421e-06  2.631381e-06
#> [281]  6.753358e-07 -6.363424e-07 -1.389599e-06  2.533580e-06 -9.268163e-07
#> [286]  8.839749e-08  3.612482e-06  8.757199e-07 -4.966487e-06 -2.869345e-09
#> [291]  4.575762e-06  2.805505e-06  4.697934e-06  4.629933e-06  2.539673e-06
#> [296] -3.205476e-06  1.547056e-06 -3.272058e-06 -3.212452e-06  2.334070e-06
#> [301]  1.405943e-06  1.431958e-07 -1.735497e-06  4.419897e-07 -2.295870e-06
#> [306]  4.036085e-07  1.899719e-06  9.654918e-07  5.237628e-07 -1.269609e-06
#> [311] -8.647822e-07 -2.569175e-06 -1.342111e-06 -1.148013e-06 -4.954353e-06
#> [316]  3.940111e-06 -1.885876e-06 -3.703507e-06 -2.591910e-06 -7.805512e-07
#> [321]  1.486158e-06  1.354874e-06 -3.414946e-07  2.816882e-06  2.937645e-06
#> [326] -5.100786e-06  4.655248e-02 -2.673122e-06  1.784683e-06  2.732874e-06
#> [331]  2.351305e-06 -6.730509e-07 -1.485266e-06 -4.476525e-07  2.010891e-07
#> [336]  1.242477e-06 -2.197624e-06  3.263849e-06 -1.109607e-06 -8.381781e-07
#> [341] -5.246660e-07  1.741311e-06  4.522540e-06  4.594098e-06 -5.867572e-03
#> [346] -4.578812e-08 -2.148003e-06  2.535349e-06  2.234590e-06  1.079569e-06
#> [351] -4.815603e-07 -1.811585e-06 -2.460988e-06  2.617199e-07 -4.231979e-06
#> [356]  7.478951e-07 -1.012135e-06  1.072562e-06  1.158262e-06  1.939598e-06
#> [361] -1.215137e-07  1.194474e-02  2.523266e-06  2.331244e-06  2.875446e-06
#> [366] -5.019541e-07 -5.312284e-06 -1.076549e-06  1.239712e-06  2.068842e-06
#> [371]  1.477681e-06 -5.083850e-06 -3.474209e-06  3.660831e-06  2.194480e-06
#> [376] -7.160934e-07 -5.324254e-06  2.345944e-06  4.523746e-07 -2.383248e-06
#> [381] -3.384738e-08 -1.097729e-02  1.059321e-06  1.104794e-06 -1.616076e-02
#> [386]  5.945619e-07 -6.224112e-08 -2.312193e-06  1.773563e-06 -1.972605e-06
#> [391] -5.604371e-06  6.310150e-07  1.531581e-03 -2.564766e-06 -1.674249e-06
#> [396] -7.453891e-08 -1.255634e-06  1.881899e-06  1.585070e-06 -2.151915e-06
#>