Inference for weighted inner product of the regression vectors in high dimensional generalized linear regressions

Inference for weighted inner product of the regression vectors in high dimensional generalized linear regressions

InnProd(
  X1,
  y1,
  X2,
  y2,
  G,
  A = NULL,
  model = c("linear", "logistic", "logistic_alter"),
  intercept = TRUE,
  beta.init1 = NULL,
  beta.init2 = NULL,
  split = TRUE,
  lambda = NULL,
  mu = NULL,
  prob.filter = 0.05,
  rescale = 1.1,
  tau = c(0.25, 0.5, 1),
  alpha = 0.05,
  verbose = FALSE
)

Arguments

X1

Design matrix for the first sample, of dimension \(n_1\) x \(p\)

y1

Outcome vector for the first sample, of length \(n_1\)

X2

Design matrix for the second sample, of dimension \(n_2\) x \(p\)

y2

Outcome vector for the second sample, of length \(n_1\)

G

The set of indices, G in the quadratic form

A

The matrix A in the quadratic form, of dimension \(|G|\times\)\(|G|\). If NULL A would be set as the \(|G|\times\)\(|G|\) submatrix of the population covariance matrix corresponding to the index set G (default = NULL)

model

The high dimensional regression model, either "linear" or "logistic" or "logistic_alter"

intercept

Should intercept(s) be fitted for the initial estimators (default = TRUE)

beta.init1

The initial estimator of the regression vector for the 1st data (default = NULL)

beta.init2

The initial estimator of the regression vector for the 2nd data (default = NULL)

split

Sampling splitting or not for computing the initial estimators. It take effects only when beta.init1 = NULL or beta.init2 = NULL. (default = TRUE)

lambda

The tuning parameter in fitting initial model. If NULL, it will be picked by cross-validation. (default = NULL)

mu

The dual tuning parameter used in the construction of the projection direction. If NULL it will be searched automatically. (default = NULL)

prob.filter

The threshold of estimated probabilities for filtering observations in logistic regression. (default = 0.05)

rescale

The factor to enlarge the standard error to account for the finite sample bias. (default = 1.1)

tau

The enlargement factor for asymptotic variance of the bias-corrected estimator to handle super-efficiency. It allows for a scalar or vector. (default = c(0.25,0.5, 1))

alpha

Level of significance to construct two-sided confidence interval (default = 0.05)

verbose

Should intermediate message(s) be printed. (default = FALSE)

Value

est.plugin

The plugin(biased) estimator for the inner product form of the regression vectors restricted to G

est.debias

The bias-corrected estimator of the inner product form of the regression vectors

se

Standard errors of the bias-corrected estimator, length of tau; corrsponding to different values of tau

ci.mat

The matrix of two.sided confidence interval for the inner product form of the regression vector; row corresponds to different values of tau

Examples

X1 = matrix(rnorm(100*5), nrow=100, ncol=5)
y1 = -0.5 + X1[,1] * 0.5 + X1[,2] * 1 + rnorm(100)
X2 = matrix(rnorm(90*5), nrow=90, ncol=5)
y2 = -0.4 + X2[,1] * 0.48 + X2[,2] * 1.1 + rnorm(90)
G = c(1,2)
A = matrix(c(1.5, 0.8, 0.8, 1.5), nrow=2, ncol=2)
Est = InnProd(X1, y1, X2, y2, G, A, model="linear")
Est$ci
#>            lower    upper
#> tau0.25 1.582242 4.829801
#> tau0.5  1.509198 4.902844
#> tau1    1.363111 5.048931