Inference for linear functional in the high dimensional linear regression

Computes the bias corrected estimator of linear functional loading\(^{\top}\beta\) for the high dimensional linear regression \(Y_i=X_i^{\top}\beta + \epsilon_i\) and the corresponding standard error.

LF(
  X,
  y,
  loading,
  intercept = TRUE,
  init.Lasso = NULL,
  lambda = NULL,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)

Arguments

X
Design matrix, of dimension \(n\) x \(p\)

y
Outcome vector, of length \(n\)

loading
Loading, of length \(p\)

intercept
Should intercept(s) be fitted (default = TRUE)

init.Lasso
Initial LASSO estimator of the regression vector (default = NULL)

lambda
The tuning parameter in the construction of LASSO estimator of the regression vector (default = NULL)

mu
The dual tuning parameter used in the construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu such that the dual optimization problem for constructing the projection direction converges (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu such that the dual optimization problem for constructing the projection direction converges (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu such that the dual optimization problem for constructing the projection direction converges (default = 10)

Value

prop.est

The bias-corrected estimator for the linear functional

se

The standard error of the bias-corrected estimator

proj

The projection direction, of length \(p\)

plug.in

The plug-in LASSO estimator for the linear functional

References

Cai T, Cai TT, Guo Z (2019). “Individualized Treatment Selection: An Optimal Hypothesis Testing Approach In High-dimensional Models.” Unknown. https://arxiv.org/pdf/1904.12891.pdf.

Examples

n = 100
p = 400
A1gen <- function(rho,p){
A1=matrix(0,p,p)
for(i in 1:p){
 for(j in 1:p){
   A1[i,j]<-rho^(abs(i-j))
 }
}
A1
}
mu <- rep(0,p)
mu[1:5] <- c(1:5)/5
rho = 0.5
Cov <- (A1gen(rho,p))/2
Cov2<-matrix(NA,nrow=p,ncol=p)
for(i in 1:p){
 for(j in 1:p){
   Cov2[i,j]<-0.5^(1+abs(i-j))
  }
}
beta <- rep(0,p)
beta[1:10] <- c(1:10)/5
X <- MASS::mvrnorm(n,mu,Cov)
y = X%*%beta + rnorm(n)
loading <- MASS::mvrnorm(1,rep(0,p),Cov2)
LF(X = X, y = y, loading = loading, intercept = TRUE)

#> [1] 3
#> [1] "step is 3"

#> $prop.est
#>        [,1]
#> [1,] 6.4332
#> 
#> $se
#> [1] 1.50963
#> 
#> $proj
#>   [1]  1.130502e-01 -2.188727e-02  5.229359e-02  5.270412e-02  2.749797e-02
#>   [6]  3.816469e-05  3.775536e-02 -2.403163e-01  4.895269e-02  8.913664e-02
#>  [11]  9.020497e-02  7.560907e-02  1.007740e-01  3.875075e-02 -5.390276e-02
#>  [16] -7.970684e-02 -2.440072e-02 -7.062679e-02  5.902480e-02  3.299823e-02
#>  [21]  3.521978e-02  3.062264e-02  9.721995e-02  1.082019e-01 -1.164647e-02
#>  [26]  9.341950e-02  4.821411e-02  3.754167e-02 -2.137850e-01 -1.246563e-01
#>  [31] -1.879916e-01 -1.089773e-01  4.490262e-02  8.169569e-02  1.176186e-01
#>  [36]  1.571845e-01  1.095737e-01  6.476010e-02  1.676349e-01  1.684602e-01
#>  [41]  1.095333e-01  1.350581e-03 -4.503489e-02 -8.742938e-02 -1.121192e-01
#>  [46] -9.243136e-02  4.890637e-02 -1.007881e-02 -6.619674e-02 -1.385545e-03
#>  [51]  6.615744e-02  4.662073e-02  6.250734e-02  9.218485e-03  4.743453e-02
#>  [56]  3.895235e-03  1.115149e-02 -1.616113e-02  2.322474e-02 -3.279609e-01
#>  [61]  8.127437e-02  5.468972e-02  4.314389e-02 -5.135382e-02 -8.568844e-02
#>  [66] -6.720111e-02 -8.638641e-02 -4.405803e-03  4.632493e-02  1.150047e-02
#>  [71]  1.129817e-02 -5.751364e-02 -1.564740e-01 -5.054604e-02 -4.761746e-02
#>  [76] -4.108234e-02 -1.348936e-02 -9.797986e-02 -4.485930e-02 -5.903388e-02
#>  [81] -2.626494e-02 -7.235988e-02  2.924883e-03 -2.346651e-02 -3.015013e-02
#>  [86] -8.646238e-02  4.149625e-02  7.245922e-02  1.469229e-01  9.547705e-02
#>  [91]  1.858358e-02 -3.878955e-01 -4.518137e-02 -2.207403e-02  8.497823e-02
#>  [96] -9.243506e-03  3.046039e-02  1.340304e-02  1.991222e-01  1.227866e-01
#> [101]  4.686110e-02 -1.738899e-02  2.331838e-02  1.961395e-02 -2.574325e-02
#> [106] -1.828194e-01 -1.566906e-02  1.362896e-02 -4.180583e-02 -1.365003e-02
#> [111]  9.265595e-02  6.972282e-02  1.650269e-02 -6.134959e-02 -5.232990e-02
#> [116] -1.293528e-01 -1.239441e-01 -1.795043e-01 -1.187249e-02 -1.257970e-01
#> [121] -3.045151e-03  5.466427e-02 -4.299975e-02 -3.691201e-02 -6.784849e-02
#> [126]  7.074773e-02  1.003652e-01  5.232467e-02 -1.023444e-02  5.104052e-02
#> [131] -9.665409e-02 -6.579667e-02 -4.007614e-02 -4.803132e-02 -4.630469e-02
#> [136] -4.544238e-02  2.043283e-02  8.206609e-03 -1.373085e-02  7.295018e-02
#> [141]  6.206388e-02 -5.303496e-02 -5.808796e-02  6.216423e-04 -5.040268e-02
#> [146]  3.346141e-02 -6.553261e-02  2.255983e-02 -3.243484e-02 -4.943406e-02
#> [151] -1.323365e-01 -2.057317e-02  3.420216e-02 -9.573373e-03 -2.348857e-02
#> [156]  4.949109e-02  5.618034e-03  1.052838e-02 -2.276923e-02 -1.176372e-02
#> [161]  2.155423e-02  5.035097e-02 -2.958981e-02  7.455506e-02  4.928766e-02
#> [166] -4.794917e-02  4.305455e-02 -6.267075e-02 -1.203605e-01 -1.198500e-01
#> [171] -5.074369e-02  4.439196e-02  9.939006e-02  5.138827e-02 -2.915189e-02
#> [176] -3.652907e-03 -9.839952e-02 -1.311371e-01 -8.720272e-02 -1.596819e-02
#> [181] -1.767465e-02 -2.356250e-02  1.423058e-01  9.876121e-02  2.023903e-02
#> [186]  1.296019e-01  1.276852e-01 -3.662324e-03 -6.042441e-02 -5.570008e-02
#> [191] -9.489476e-04 -1.180589e-01 -2.720429e-02 -4.379431e-02 -1.475769e-01
#> [196]  1.882983e-02 -1.048159e-01 -9.086699e-02 -1.340231e-01  5.903651e-02
#> [201]  9.535273e-02  3.664637e-02 -5.546842e-02 -4.946596e-02  1.104612e-02
#> [206] -5.971581e-02  1.468310e-01  1.268229e-01  7.271401e-02  1.057888e-01
#> [211]  6.440960e-02 -1.556125e-01 -9.664643e-02 -2.834159e-02  1.587973e-01
#> [216]  2.182405e-01  3.866117e-02 -5.058659e-03 -6.737094e-05 -8.382575e-02
#> [221] -6.715282e-02 -1.154548e-01 -4.866029e-02  3.453456e-02 -1.221533e-03
#> [226]  8.029684e-02  4.885684e-02  9.882568e-02  8.144807e-03  4.121638e-02
#> [231]  1.321617e-01  9.164720e-02  2.734275e-01  1.155656e-01  2.245564e-01
#> [236]  9.788975e-03  1.386244e-02  4.000412e-02  3.153705e-02 -1.342295e-01
#> [241] -3.734939e-02 -9.155645e-02 -8.502146e-02  2.538197e-02  1.297496e-02
#> [246]  7.746646e-02  2.668405e-02 -1.586606e-01 -2.455740e-01 -1.433727e-01
#> [251] -4.986431e-02 -4.773619e-02  6.700333e-02  6.020406e-03  4.594531e-02
#> [256]  8.321287e-02 -7.345803e-02 -1.772243e-01 -2.076814e-01 -1.140920e-01
#> [261] -4.233382e-02 -5.497605e-02  3.143334e-02  2.789148e-02  3.824176e-02
#> [266]  3.958670e-02  7.144549e-02 -3.388763e-02 -1.061599e-01 -1.192311e-01
#> [271] -6.336736e-02 -7.403462e-02  7.608271e-02  7.617826e-04  1.225105e-01
#> [276]  1.947855e-01 -6.882982e-03  1.296173e-02  2.649909e-02  7.854022e-02
#> [281]  1.097824e-01  3.435123e-01  1.278670e-01 -9.454828e-02 -3.549969e-02
#> [286] -1.191202e-01  2.704304e-01 -3.714909e-02 -5.965492e-02  1.115179e-02
#> [291]  3.920364e-02  5.369127e-02  3.992554e-02  3.295731e-02 -6.192887e-02
#> [296] -1.906290e-03  1.027402e-01  7.875947e-02  1.658419e-01 -2.518852e-02
#> [301]  1.801687e-01  6.367642e-02 -3.330569e-02 -8.289172e-02 -4.320864e-02
#> [306] -3.961617e-02 -1.280889e-01 -2.123027e-02 -1.144350e-01 -1.093114e-01
#> [311]  1.398180e-02  8.428887e-02  6.560807e-02 -1.192685e-01 -3.025728e-01
#> [316] -1.070811e-01 -1.195211e-01  7.494600e-02 -1.958308e-02  3.219249e-02
#> [321]  1.506674e-01  1.531688e-01  8.713382e-02  9.624283e-02 -2.776897e-02
#> [326]  2.208440e-02 -2.359146e-02  3.099839e-02 -1.450363e-02  4.013212e-02
#> [331]  7.652855e-02  1.969862e-01  2.306232e-01  3.624671e-02 -2.086142e-02
#> [336] -5.993315e-02  2.064305e-02 -1.058442e-01 -6.647455e-02 -1.227112e-01
#> [341] -7.600058e-02 -1.197372e-01 -1.154656e-01 -1.199773e-02  4.443043e-02
#> [346] -1.282035e-02 -7.401473e-02 -2.809404e-03  1.776912e-01  8.862438e-02
#> [351]  2.875638e-02  7.936341e-02  1.024469e-01  4.666990e-02  4.395139e-02
#> [356]  5.898913e-02  4.252838e-02  3.567832e-02  1.758626e-01  1.626727e-01
#> [361]  1.129815e-01 -8.315755e-02 -9.018081e-03 -8.721113e-02 -4.967281e-02
#> [366]  3.342401e-02  8.552305e-03  2.100803e-01  1.914683e-01  3.751005e-02
#> [371]  8.699636e-02 -2.207720e-02  6.078786e-02  3.835914e-02  1.195336e-01
#> [376] -2.261174e-02  9.933915e-02  1.731681e-01  1.114819e-01  4.897603e-02
#> [381]  5.218957e-02  4.265298e-02  1.281942e-01  9.344387e-02  1.629630e-02
#> [386]  8.123778e-02  9.782051e-02  2.510390e-01  1.434584e-01 -1.549435e-01
#> [391] -3.432072e-02 -9.536287e-03  3.704661e-02  2.244580e-02  1.074853e-02
#> [396] -6.206502e-02 -2.254728e-02  1.902413e-02  3.480277e-02 -1.324308e-02
#> [401]  4.689829e-02
#> 
#> $plug.in
#> [1] 3.248337
#>