Inference for quadratic forms of the regression vector in high dimensional generalized linear regressions

Inference for quadratic forms of the regression vector in high dimensional generalized linear regressions

Usage

QF(
  X,
  y,
  G,
  A = NULL,
  model = c("linear", "logistic", "logistic_alter"),
  intercept = TRUE,
  beta.init = NULL,
  split = TRUE,
  lambda = NULL,
  mu = NULL,
  prob.filter = 0.05,
  rescale = 1.1,
  tau = c(0.25, 0.5, 1),
  alpha = 0.05,
  verbose = FALSE
)

Arguments

X

Design matrix, of dimension \(n\) x \(p\)

y

Outcome vector, of length \(n\)

G

The set of indices, G in the quadratic form

A

The matrix A in the quadratic form, of dimension \(|G|\times\)\(|G|\). If NULL A would be set as the \(|G|\times\)\(|G|\) submatrix of the population covariance matrix corresponding to the index set G (default = NULL)

model

The high dimensional regression model, either "linear" or "logistic" or "logistic_alter"

intercept

Should intercept be fitted for the initial estimator (default = TRUE)

beta.init

The initial estimator of the regression vector (default = NULL)

split

Sampling splitting or not for computing the initial estimator. It take effects only when beta.init = NULL. (default = TRUE)

lambda

The tuning parameter in fitting initial model. If NULL, it will be picked by cross-validation. (default = NULL)

mu

The dual tuning parameter used in the construction of the projection direction. If NULL it will be searched automatically. (default = NULL)

prob.filter

The threshold of estimated probabilities for filtering observations in logistic regression. (default = 0.05)

rescale

The factor to enlarge the standard error to account for the finite sample bias. (default = 1.1)

tau

The enlargement factor for asymptotic variance of the bias-corrected estimator to handle super-efficiency. It allows for a scalar or vector. (default = c(0.25,0.5,1))

alpha

Level of significance to construct two-sided confidence interval (default = 0.05)

verbose

Should intermediate message(s) be printed. (default = FALSE)

Value

est.plugin

The plugin(biased) estimator for the quadratic form of the regression vector restricted to G

est.debias

The bias-corrected estimator of the quadratic form of the regression vector

se

Standard errors of the bias-corrected estimator, length of tau; corrsponding to different values of tau

ci.mat

The matrix of two.sided confidence interval for the quadratic form of the regression vector; row corresponds to different values of tau

Examples

X = matrix(rnorm(100*5), nrow=100, ncol=5)
y = X[,1] * 0.5 + X[,2] * 1 + rnorm(100)
G = c(1,2)
A = matrix(c(1.5, 0.8, 0.8, 1.5), nrow=2, ncol=2)
Est = QF(X, y, G, A, model="linear")
## compute confidence intervals
ci(Est, alpha=0.05, alternative="two.sided")
#>    tau    lower    upper
#> 1 0.25 1.135152 4.720540
#> 2 0.50 1.129802 4.725889
#> 3 1.00 1.119152 4.736540

## summary statistics
summary(Est)
#> Call: 
#> Inference for Quadratic Functional
#> 
#>   tau est.plugin est.debias Std. Error z value Pr(>|z|)   
#>  0.25      3.425      2.928     0.9147   3.201 0.001369 **
#>  0.50      3.425      2.928     0.9174   3.192 0.001415 **
#>  1.00      3.425      2.928     0.9228   3.173 0.001510 **